Nifty restructuration
