+++ Firmware/CoverUI/platformio.ini
@@ -303,7 +303,7 @@ debug_server =
 board = disco_f030r8
 build_flags =
   ${env.build_flags}
-  -include YardForce/include/MDL_RM-ECOW-V100.h
+  -include YardForce/MDL_RM-ECOW-V100.h
   -DMCU_STM32
 
 upload_flags =
